add Tmp and Item, finish Email validator
Tmp: some kind of sandbox - does something with with the value, but
doesn't return the result.
Item: gets an item from a list or dict. If a validator is given,
it writes the result back to the list/dict.

since Email is working, we've also got a working Schema, fieldset,
pre and post_validate.